requirements.txt
boto3==1.26.26
botocore==1.29.26
jmespath==1.0.1
numpy==1.23.5
pandas==1.5.2
psycopg2-binary==2.9.5
python-dateutil==2.8.2
pytz==2022.6
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.13
